Extract content controller into own namespace
