+++ config/routes.rb
@@ -1,19 +1,20 @@
 Rails.application.routes.draw do
+  namespace :content do
+    mount Ckeditor::Engine => '/ckeditor'
+    devise_for :users, class_name: 'Content::User'
+    ActiveAdmin.routes(self)
+
+    get '/posts', to: 'content#posts'
+    get '/site_dive/:id', to: 'content#site_dive'
+    get '/tweets', to: 'content#tweets'
+  end
+
   namespace :api do
     namespace :v3 do
       get '/get_contexts', to: 'context#get_contexts'
       get '/get_all_nodes', to: 'nodes#get_all_nodes'
     end
     namespace :v2 do
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-      mount Ckeditor::Engine => '/ckeditor'
-      devise_for :users, class_name: 'Content::User'
-      ActiveAdmin.routes(self)
-
-      get '/posts', to: 'content#posts'
-      get '/site_dive/:id', to: 'content#site_dive'
-      get '/tweets', to: 'content#tweets'
-
       resources :geo_id, only: :index
       resources :download, only: [:index], as: :download
       resources :indicators, only: [:index]
