+++ spec/responses/posts_spec.rb
@@ -3,9 +3,9 @@ require 'rails_helper'
 RSpec.describe 'Get posts', type: :request do
   include_context 'posts'
 
-  describe 'GET /api/v2/posts' do
+  describe 'GET /content/posts' do
     it 'has the correct response structure' do
-      get '/api/v2/posts'
+      get '/content/posts'
 
       expect(@response.status).to eq 200
       expect(@response).to match_response_schema('posts')
